schedule_manage/requirements.txt
# requirements.txt
# ===================================
# תלויות מערכת השיבוץ החכמה
# ===================================

# Django Framework
Django>=4.2.0,<5.0
djangorestframework>=3.14.0

# מסד נתונים
psycopg2-binary>=2.9.0  # PostgreSQL
# או
# mysqlclient>=2.1.0    # MySQL (אלטרנטיבה)

# OR-Tools - מנוע האופטימיזציה הראשי
ortools>=9.7.2996

# עיבוד נתונים ומתמטיקה
numpy>=1.21.0
pandas>=1.5.0
scipy>=1.9.0

# ייצוא לאקסל
openpyxl>=3.0.0
xlsxwriter>=3.0.0

# גרפים ויזואליזציה (אופציונלי)
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# עיבוד תאריכים מתקדם
python-dateutil>=2.8.0

# לוגינג מתקדם
structlog>=22.0.0

# API ותיעוד
drf-spectacular>=0.25.0  # OpenAPI/Swagger
django-cors-headers>=3.13.0
django-filter>=22.1

# אבטחה
django-ratelimit>=3.0.0
cryptography>=3.4.0

# ביצועים וזיכרון
memory-profiler>=0.60.0
psutil>=5.8.0

# בדיקות
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
factory-boy>=3.2.0
freezegun>=1.2.0

# Development Tools
black>=22.0.0           # Code formatting
flake8>=5.0.0          # Linting
mypy>=0.991            # Type checking
pre-commit>=2.20.0     # Git hooks

# רקע וזמן אמת (אופציונלי)
celery>=5.2.0          # Background tasks
redis>=4.3.0           # Caching & message broker

# Monitoring ולוגים (production)
sentry-sdk>=1.9.0      # Error tracking
django-extensions>=3.2.0

# אבטחה נוספת (production)
django-security>=0.17.0
django-csp>=3.7

# ===================================
# הערות התקנה:
# ===================================

# התקנה בסיסית:
# pip install -r requirements.txt

# התקנה לפיתוח (כולל כלי פיתוח):
# pip install -r requirements.txt
# pip install -r requirements-dev.txt

# התקנה לייצור:
# pip install -r requirements.txt --no-dev

# עבור מערכות עם מגבלות זיכרון:
# pip install --no-cache-dir -r requirements.txt

# עבור Apple Silicon (M1/M2):
# ARCHFLAGS="-arch arm64" pip install -r requirements.txt

# ===================================
# תלויות מערכת (system dependencies):
# ===================================

# Ubuntu/Debian:
# sudo apt-get update
# sudo apt-get install python3-dev libpq-dev build-essential

# CentOS/RHEL:
# sudo yum install python3-devel postgresql-devel gcc

# macOS:
# brew install postgresql
# xcode-select --install

# Windows:
# מומלץ להשתמש ב-conda או להתקין Visual Studio Build Tools

# ===================================
# הגדרות סביבת פיתוח מומלצות:
# ===================================

# יצירת סביבה וירטואלית:
# python -m venv scheduling_env
# source scheduling_env/bin/activate  # Linux/Mac
# scheduling_env\Scripts\activate     # Windows

# עדכון pip:
# pip install --upgrade pip setuptools wheel

# התקנה מהירה (ללא חבילות פיתוח):
# pip install ortools pandas numpy openpyxl Django djangorestframework

# בדיקת התקנה:
# python -c "import ortools; print('OR-Tools version:', ortools.__version__)"
# python -c "import pandas; print('Pandas version:', pandas.__version__)"
# python -c "import numpy; print('NumPy version:', numpy.__version__)"

# ===================================
# חבילות אופציונליות לפי צרכים:
# ===================================

# לעבודה עם קבצי Excel מורכבים:
# xlrd>=2.0.0
# xlwt>=1.3.0

# לעבודה עם CSV בעברית:
# chardet>=5.0.0

# לביצועים מתקדמים:
# numba>=0.56.0
# cython>=0.29.0

# לעבודה עם API חיצוניים:
# requests>=2.28.0
# httpx>=0.23.0

# לעבודה עם זמן אמת:
# django-channels>=4.0.0
# channels-redis>=4.0.0

# ===================================
# גרסאות נבדקות ומומלצות:
# ===================================

# Python: 3.9, 3.10, 3.11 (מומלץ: 3.10)
# Django: 4.2 LTS (מומלץ)
# OR-Tools: 9.7+ (מומלץ: הגרסה האחרונה)

# ===================================
# בעיות נפוצות ופתרונות:
# ===================================

# בעיה: OR-Tools לא מתקין
# פתרון: pip install --upgrade pip && pip install ortools

# בעיה: שגיאות compilation
# פתרון: התקינו build tools והתלויות המערכת

# בעיה: שגיאות זיכרון בשיבוצים גדולים  
# פתרון: הגדלת זיכרון או שימוש בפרמטרים מותאמים

# בעיה: ביצועים איטיים
# פתרון: בדקו שיש מספיק CPU cores זמינים לOR-Tools
